requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
playwright>=1.52.0
